REGISTRY-3624 - Improved the unbounded filed search using a concurrent hash map and refactored the code.
