+++ components/registry/org.wso2.carbon.registry.indexing/src/main/java/org/wso2/carbon/registry/indexing/service/RxtUnboundedFieldManagerService.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.carbon.registry.indexing;
+package org.wso2.carbon.registry.indexing.service;
 
 import org.wso2.carbon.context.PrivilegedCarbonContext;
 import org.wso2.carbon.core.AbstractAdmin;
@@ -24,23 +24,22 @@ import org.wso2.carbon.registry.core.exceptions.RegistryException;
 import org.wso2.carbon.registry.core.session.UserRegistry;
 import org.wso2.carbon.registry.indexing.internal.RxtDataServiceDataHolder;
 import org.wso2.carbon.registry.indexing.utils.RxtUnboundedDataLoadUtils;
-import org.wso2.carbon.user.api.Tenant;
-import org.wso2.carbon.user.api.TenantManager;
-import org.wso2.carbon.user.api.UserStoreException;
 
-import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
 
 /**
  * This class is used to set rxt unbounded filed details to the memory.
  */
-public class RxtUnboundedFiledManager extends AbstractAdmin {
+public class RxtUnboundedFieldManagerService extends AbstractAdmin {
 
-    private static RxtUnboundedFiledManager rxtUnboundedFiledManagerInstance = new RxtUnboundedFiledManager();
-    private static HashMap<Integer, HashMap<String, List<String>>> allTenantsUnboundedFields = new HashMap<>();
+    private static RxtUnboundedFieldManagerService rxtUnboundedFieldManagerServiceInstance
+            = new RxtUnboundedFieldManagerService();
+    private static Map<Integer, Map<String, List<String>>> allTenantsUnboundedFields = new ConcurrentHashMap<>();
 
-    public static RxtUnboundedFiledManager getInstance() {
-        return rxtUnboundedFiledManagerInstance;
+    public static RxtUnboundedFieldManagerService getInstance() {
+        return rxtUnboundedFieldManagerServiceInstance;
     }
 
     /**
@@ -48,7 +47,7 @@ public class RxtUnboundedFiledManager extends AbstractAdmin {
      *
      * @return unbounded rxt filed values.
      */
-    public HashMap<Integer, HashMap<String, List<String>>> getTenantsUnboundedFileds() {
+    public Map<Integer, Map<String, List<String>>> getTenantsUnboundedFields() {
         return allTenantsUnboundedFields;
     }
 
@@ -62,17 +61,17 @@ public class RxtUnboundedFiledManager extends AbstractAdmin {
 
         int tenantId = PrivilegedCarbonContext.getThreadLocalCarbonContext().getTenantId();
         UserRegistry registry = RxtDataServiceDataHolder.getInstance().getRegistryService().getRegistry();
-        HashMap<String, List<String>> tenantRxtUnboundedEntries = RxtUnboundedDataLoadUtils.getRxtData(registry);
+        Map<String, List<String>> tenantRxtUnboundedEntries = RxtUnboundedDataLoadUtils.getRxtData(registry);
         allTenantsUnboundedFields.put(tenantId, tenantRxtUnboundedEntries);
     }
 
     /**
      * This method is used to set a specific tenants unbounded filed values.
      *
-     * @param tenantId      tenant Id
-     * @param rxtUnboundedFiledMap    rxt unbounded filed map.
+     * @param tenantId             tenant Id
+     * @param rxtUnboundedFiledMap rxt unbounded filed map.
      */
-    public void setTenantsUnboundedFields(Integer tenantId, HashMap<String, List<String>> rxtUnboundedFiledMap){
+    public void setTenantsUnboundedFields(Integer tenantId, Map<String, List<String>> rxtUnboundedFiledMap) {
         allTenantsUnboundedFields.put(tenantId, rxtUnboundedFiledMap);
     }
 }
