+++ components/registry/org.wso2.carbon.registry.indexing/src/main/java/org/wso2/carbon/registry/indexing/solr/SolrClient.java
@@ -45,7 +45,7 @@ import org.wso2.carbon.registry.core.session.UserRegistry;
 import org.wso2.carbon.registry.indexing.AsyncIndexer;
 import org.wso2.carbon.registry.indexing.IndexingConstants;
 import org.wso2.carbon.registry.indexing.RegistryConfigLoader;
-import org.wso2.carbon.registry.indexing.RxtUnboundedFiledManager;
+import org.wso2.carbon.registry.indexing.service.RxtUnboundedFieldManagerService;
 import org.wso2.carbon.registry.indexing.SolrConstants;
 import org.wso2.carbon.registry.indexing.Utils;
 import org.wso2.carbon.registry.indexing.indexer.Indexer;
@@ -1285,11 +1285,11 @@ public class SolrClient {
      */
     private boolean isMultiValueField(String mediaType, String fieldKey) {
         boolean result = false;
-        HashMap<Integer, HashMap<String, List<String>>> allTenantsUnboundedFields = RxtUnboundedFiledManager
-                .getInstance().getTenantsUnboundedFileds();
+        Map<Integer, Map<String, List<String>>> allTenantsUnboundedFields = RxtUnboundedFieldManagerService
+                .getInstance().getTenantsUnboundedFields();
 
         if (allTenantsUnboundedFields.size() > 0) {
-            HashMap<String, List<String>> rxtDetails = allTenantsUnboundedFields.get(PrivilegedCarbonContext
+            Map<String, List<String>> rxtDetails = allTenantsUnboundedFields.get(PrivilegedCarbonContext
                     .getThreadLocalCarbonContext().getTenantId());
             if (rxtDetails != null) {
                 List<String> fields = rxtDetails.get(mediaType);
